Move ModsTab out of ui.py
